Fix most warnings reported by sparse
Leave only the relatively hard stuff.  Most changes are just
annotations, but there are two fixes that may affect the functionality.

__constant_htons() is no longer used on a non-constant field, which
affect EAPOL frame detection.

flags in struct ieee80211_cb was extended to 32 bits to match the trunk. 
This may cause some flags to start working the way the were meant to.


git-svn-id: http://madwifi-project.org/svn/branches/madwifi-old-openhal@2046 0192ed92-7a03-0410-a25b-9323aeb14dbd