Fix several cases of inconsistent locking in ieee80211_node.c
Ingo Molnar's runtime locking correctness validator has indicated
several problems with the driver, when the same lock (or, more
precisely, a lock from the same class) is used with softirq enabled
(i.e. softirq are already enabled and the lock doesn't disable them) and
from a softirq handler.

That can result in a lockup when the code holding the lock is
interrupted by a softirq and tries to acquire the lock again.

nt_nodelock and nt_scanlock in struct ieee80211_node_table had such
problem.  In all cases, the fix was to disable softirqs when acquiring
those locks in the code that can run with softirqs enabled.

Only actually detected problems have been fixed, and only after careful
analysis.


git-svn-id: http://madwifi-project.org/svn/branches/madwifi-old-openhal@2117 0192ed92-7a03-0410-a25b-9323aeb14dbd