Fix memory corruption on PowerPC and possibly other platforms
The size of the memory window for consistent DMA must be a power of 2. 
Otherwise, it's supposed to be rounded up.  Due to a bug in Linux kernel
headers, it is rounded down for constant values.  The bug was fixed days
ago and the fix will appear in Linux 2.6.21.

Put barrier() to ensure that sc->sc_desc_len is not treated as a
constant.


git-svn-id: http://madwifi-project.org/svn/branches/madwifi-old-openhal@2198 0192ed92-7a03-0410-a25b-9323aeb14dbd