Fix WEP on big-endian systems
key_v in ath5k_hw_set_key() should be little-endian, since it's
populated by memcpy().  Copy the key in chunks of 6, not 4, as we don't
care about key_v element boundary as long as it's fixed-endian and
memcpy() is used.  Use a separate variable for key type to avoid
byteswapping it twice and to make a good use of AR5K_KEYTABLE_TYPE.

Don't use AR5K_KEYCACHE_SIZE, it's irrelevant.  We need 5 registers
because that's how many registers it takes to keep a 128-bit (16-byte)
key in 6-byte chunks with 2-byte gaps (6 + 2 + 6 + 2 + 4 = 20 bytes = 5
registers).


git-svn-id: http://madwifi-project.org/svn/branches/madwifi-old-openhal@2531 0192ed92-7a03-0410-a25b-9323aeb14dbd